requirements.txt
anapioficeandfire==0.1.2
Django==2.2.5
djangorestframework==3.10.3
mysqlclient==1.4.4
pkg-resources==0.0.0
pytz==2019.2
requests==2.9.1
six==1.10.0
sqlparse==0.3.0
ujson==1.35
